# public/json/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Defacto2 API",
    "description": "A RESTful API for accessing historical Scene data and digital culture. Provides both HTML and plain text content formats for API consumers.\n\nNote: This OpenAPI specification is completely generated by Mistral Vibe using the devstral-2 model.",
    "version": "1.0.0",
    "contact": {
      "name": "Defacto2 API Support",
      "url": "https://defacto2.net"
    },
    "license": {
      "name": "CC BY 4.0",
      "url": "https://creativecommons.org/licenses/by/4.0/"
    }
  },
  "servers": [
    {
      "url": "https://defacto2.net/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/api/v1/milestones": {
      "get": {
        "tags": ["milestones"],
        "summary": "Get all milestones",
        "description": "Returns a complete list of all historical milestones with clean HTML and plain text content.",
        "operationId": "getAllMilestones",
        "responses": {
          "200": {
            "description": "A list of milestones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Milestone"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/areacodes": {
      "get": {
        "tags": ["areacodes"],
        "summary": "Get all area codes",
        "description": "Returns a complete list of all North American Numbering Plan area codes with their associated regions and historical notes.",
        "operationId": "getAllAreacodes",
        "responses": {
          "200": {
            "description": "A list of area codes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Areacode"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/categories": {
      "get": {
        "tags": ["categories"],
        "summary": "Get all artifact categories",
        "description": "Returns a complete list of all artifact categories with their descriptions and URL mappings for API, HTML3, and HTML views.",
        "operationId": "getAllCategories",
        "responses": {
          "200": {
            "description": "A list of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platforms": {
      "get": {
        "tags": ["platforms"],
        "summary": "Get all operating system platforms",
        "description": "Returns a complete list of operating system platforms (ANSI, Audio, DOS, Windows, Linux, Java, macOS, etc.) with their descriptions and URL mappings for API, HTML3, and HTML views.",
        "operationId": "getAllPlatforms",
        "responses": {
          "200": {
            "description": "A list of platforms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Platform"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/groups": {
      "get": {
        "tags": ["artifacts"],
        "summary": "Get all releaser groups",
        "description": "Returns a paginated list of all releaser groups (collectives, individuals, brands) with their statistics and URLs. Each group includes total file count and total size information.",
        "operationId": "getAllGroups",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (1000 items per page)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of releaser groups",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "releasers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReleaserGroup"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "description": "Current page number",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "Total number of pages available",
                      "example": 4
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid page parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/magazines": {
      "get": {
        "tags": ["artifacts"],
        "summary": "Get all magazines",
        "description": "Returns a list of all scene magazines with their statistics and URLs. Each magazine includes total file count and total size information.",
        "operationId": "getAllMagazines",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (1000 items per page)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of magazines",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "releasers": {
                      "type": "array",
                      "description": "Array of magazine entries",
                      "items": {
                        "$ref": "#/components/schemas/ReleaserGroup"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "description": "Current page number",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "Total number of pages available",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid page parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/boards": {
      "get": {
        "tags": ["artifacts"],
        "summary": "Get all BBS boards",
        "description": "Returns a complete list of all BBS (Bulletin Board System) boards with their statistics and URLs. Each BBS includes total file count and total size information.",
        "operationId": "getAllBBS",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (always 1, included for API consistency)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A complete list of all BBS boards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "releasers": {
                      "type": "array",
                      "description": "Array of all BBS board entries",
                      "items": {
                        "$ref": "#/components/schemas/ReleaserGroup"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "description": "Current page number (always 1)",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "Total number of pages (always 1 for BBS)",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid page parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sites": {
      "get": {
        "tags": ["artifacts"],
        "summary": "Get all FTP sites",
        "description": "Returns a complete list of all FTP (File Transfer Protocol) sites with their statistics and URLs. Each FTP site includes total file count and total size information.",
        "operationId": "getAllSites",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (always 1, included for API consistency)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A complete list of all FTP sites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "releasers": {
                      "type": "array",
                      "description": "Array of all FTP site entries",
                      "items": {
                        "$ref": "#/components/schemas/ReleaserGroup"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "description": "Current page number (always 1)",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "Total number of pages (always 1 for FTP sites)",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid page parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts": {
      "get": {
        "tags": ["artifacts"],
        "summary": "Get all artifacts",
        "description": "Returns a paginated list of all artifacts in the database, including comprehensive metadata for each artifact. Artifacts are ordered by ID for consistent pagination.",
        "operationId": "getAllArtifacts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (1000 files per page)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifacts": {
                      "type": "array",
                      "description": "Array of file entries",
                      "items": {
                        "$ref": "#/components/schemas/FileRecord"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "description": "Current page number",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "Total number of pages available",
                      "example": 42
                    },
                    "totalFiles": {
                      "type": "integer",
                      "description": "Total number of files in the database",
                      "example": 42000
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid page parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/artifacts/new": {
      "get": {
        "tags": ["artifacts"],
        "summary": "Get new artifacts",
        "description": "Returns a paginated list of recently added artifacts (new uploads), ordered by ID (newest first). This endpoint provides the same data as /api/artifacts but with semantic meaning for new uploads.",
        "operationId": "getNewArtifacts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "example": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of files per page (fixed at 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1000,
              "example": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of new uploads",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifacts": {
                      "type": "array",
                      "description": "Array of new file entries",
                      "items": {
                        "$ref": "#/components/schemas/ArtifactSum"
                      }
                    },
                    "statistics": {
                      "type": "object",
                      "description": "Overall statistics for all files",
                      "properties": {
                        "totalFiles": {
                          "type": "integer",
                          "description": "Total number of files in the database",
                          "example": 10000
                        },
                        "totalSize": {
                          "type": "string",
                          "description": "Total size of all files (human-readable)",
                          "example": "10 GB"
                        },
                        "totalSizeBytes": {
                          "type": "integer",
                          "description": "Total size of all files in bytes",
                          "example": 10737418240
                        }
                      }
                    },
                    "page": {
                      "type": "integer",
                      "description": "Current page number",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "Total number of pages available",
                      "example": 500
                    },
                    "limit": {
                      "type": "integer",
                      "description": "Number of files per page",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid page or limit parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demozoo": {
      "get": {
        "tags": ["demozoo"],
        "summary": "Get all Demozoo group mappings",
        "description": "Returns a complete list of all groups that have Demozoo entries with their URIs, IDs, and direct URLs to their Demozoo group pages. This endpoint is useful for cross-referencing Defacto2 groups with Demozoo entries.",
        "operationId": "getDemozooGroups",
        "responses": {
          "200": {
            "description": "A list of groups with Demozoo mappings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "groups": {
                      "type": "array",
                      "description": "Array of group mappings",
                      "items": {
                        "$ref": "#/components/schemas/DemozooGroup"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Total number of groups with Demozoo entries"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/websites": {
      "get": {
        "tags": ["websites"],
        "summary": "Get all curated websites",
        "description": "Returns a complete alphabetical list of all curated websites about the Scene and digital culture, including text art, BBS, pirate scene, demoscene, YouTube videos, podcasts, books, documentaries, and AMAs.",
        "operationId": "getAllWebsites",
        "responses": {
          "200": {
            "description": "A list of websites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "websites": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Website"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Total number of websites"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Milestone": {
        "type": "object",
        "properties": {
          "Year": {
            "type": "integer",
            "description": "The year the milestone occurred",
            "example": 1971
          },
          "Month": {
            "type": "integer",
            "description": "The month the milestone occurred (1-12)",
            "example": 10,
            "minimum": 1,
            "maximum": 12
          },
          "Day": {
            "type": "integer",
            "description": "The day the milestone occurred",
            "example": 15,
            "minimum": 1,
            "maximum": 31
          },
          "Title": {
            "type": "string",
            "description": "The title or name of the milestone",
            "example": "Secrets of the Little Blue Box"
          },
          "Lead": {
            "type": "string",
            "description": "Short introductory text or subtitle",
            "example": "Esquire October 1971"
          },
          "Content": {
            "type": "string",
            "description": "Plain text version of the content (all HTML tags removed)",
            "example": "Ron Rosenbaum writes the first mainstream article on phone freaks..."
          },
          "ContentHTML": {
            "type": "string",
            "description": "Main content with clean HTML (presentation attributes removed)",
            "example": "<p>Ron Rosenbaum writes the first mainstream article on phone freaks...</p>"
          },
          "Link": {
            "type": "string",
            "description": "URL to an article or external resource about the milestone",
            "example": "https://www.slate.com/articles/technology/the_spectator/2011/10/the_article_that_inspired_steve_jobs_secrets_of_the_little_blue_.html"
          },
          "LinkTitle": {
            "type": "string",
            "description": "Title for the external link",
            "example": "the complete article"
          },
          "Highlight": {
            "type": "boolean",
            "description": "Whether this milestone is marked as significant/highlighted",
            "example": true
          },
          "List": {
            "type": "array",
            "description": "Additional related links",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          },
          "Picture": {
            "$ref": "#/components/schemas/Picture"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid year format"
          }
        }
      },
      "Areacode": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Three-digit North American Numbering Plan area code",
            "example": 212,
            "minimum": 200,
            "maximum": 999
          },
          "territories": {
            "type": "array",
            "description": "List of territories (states, provinces) that use this area code",
            "items": {
              "type": "string"
            },
            "example": ["New York"]
          },
          "notes": {
            "type": "string",
            "description": "Historical notes about the area code (splits, creation dates, etc.)",
            "example": "New York City, but after December 1984 it split to 212 (Manhattan and the Bronx) and 718 (Brooklyn, Queens, and Staten Island)."
          }
        }
      },
...
    }
  }
}